fastapi-backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.23.0
pydantic>=2.4.0
python-multipart>=0.0.6

# Azure OpenAI
openai>=1.3.0

# Environment management
python-dotenv>=1.0.0

# Diagram generation
diagrams>=0.23.4
graphviz>=0.20.1

# Development dependencies (optional)
pytest>=7.4.0
httpx>=0.25.0